+++ frontend/components/page/head.php
@@ -4,6 +4,10 @@
 <?php $styles = array_unique( array_merge($styles, $default_styles) ) ?>
 
 <head>
+    <meta http-equiv="Content-Type" content="text/html;" charset="utf-8" /> 
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
     <?= c('page/seo', [ 'title' => $title ]) ?>
     <style>
         <?php foreach ($styles as $style): ?>