+++ frontend/components/page/seo.php
@@ -45,7 +45,4 @@
 
 <title><?= $title ?></title>
 <meta name="description" content="<?= APP_DESC ?>">
-<meta charset="UTF-8">
-<meta http-equiv="X-UA-Compatible" content="IE=edge">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
 <link rel="shortcut icon" href="<?= $cover_img ?>" type="image/x-icon">